move output create to its own file
